Persistence in a new file
I moved the functions related to persistence to a file called
persistence.lua.
